♻️ Refactor: useQuery 사용을 api 페이지롤 분리
